requirements.txt
tweepy
pillow
python-dateutil